categories list - refactoring, moved categories list and associated files to seprate folder.
