+++ frontend/src/components/Home/CategoriesList/CategoriesList.js
