+++ frontend/src/components/Home/CategoriesList/actions.js
