+++ frontend/src/components/Home/CategoriesList/constants.js
