+++ frontend/src/rootReducer/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import categories from '../components/Home/reducers';
+import categories from '../components/Home/CategoriesList/reducers';
 
 export default combineReducers({
   categories
